backend/requirements.txt
autogen-agentchat>=0.3.0
autogen-ext>=0.3.0
openai>=1.40.0
cerebras-cloud-sdk>=0.1.2
faiss-cpu>=1.8.0
sqlalchemy>=2.0.30
trimesh>=4.3.0
pygltflib>=1.16.5
shapely>=2.0.4
opensimplex>=0.4.5
numpy>=1.26
pandas>=2.2
python-dotenv>=1.0
pillow>=10.3
uvicorn>=0.30
fastapi>=0.110
rich>=13.7
google-genai>=1.40
